services/mnemosyne/pyproject.toml
[tool.poetry]
name = "mnemosyne"
version = "0.1.0"
description = "Mnemosyne - memória viva do ecossistema"
authors = ["Atlas Platform <noreply@example.com>"]
package-mode = true
readme = "README.md"
packages = [{include = "mnemosyne", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pydantic = "^2.7.0"
pymongo = "^4.7.2"
boto3 = "^1.34.69"
opentelemetry-api = "^1.24.0"
opentelemetry-sdk = "^1.24.0"
opentelemetry-exporter-otlp = "^1.24.0"
opentelemetry-instrumentation-fastapi = "^0.45b0"
uvicorn = {version = "^0.30.0", optional = true}

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
